fixed more warnings, added some js bindings, added some missing initializers
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@21471 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ y60/jsgtk/JSRenderArea.cpp
@@ -145,7 +145,7 @@ JSRenderArea::getPropertySwitch(NATIVE & theNative, unsigned long theID,
     switch (theID) {
         case 0:
         default:
-            if (theID >= JSGTKBASE::PROP_BEGIN) {
+            if (theID >= static_cast<unsigned long>(JSGTKBASE::PROP_BEGIN)) {
                 return JSGTKBASE::getPropertySwitch(theNative, theID, cx, obj, id, vp);
             } else {
                 return JSY60BASE::getPropertySwitch(theNative, theID, cx, obj, id, vp);
@@ -159,7 +159,7 @@ JSRenderArea::setPropertySwitch(NATIVE & theNative, unsigned long theID,
     switch (theID) {
         case 0:
         default:
-            if (theID >= JSGTKBASE::PROP_BEGIN) {
+            if (theID >= static_cast<unsigned long>(JSGTKBASE::PROP_BEGIN)) {
                 return JSGTKBASE::setPropertySwitch(theNative, theID, cx, obj, id, vp);
             } else {
                 return JSY60BASE::setPropertySwitch(theNative, theID, cx, obj, id, vp);
